requirements.txt
git+https://github.com/Rapptz/discord.py
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
aiohttp
bs4
python-dotenv
